Added function to update fn121 ProcessType and add basename to msg
This commit updates the append to nearshore functions so that process
type is cacluated properly based on the number of records in the FN122
table and the associated gear type. It also updates the readme message
to include the name of the source database (rather than a simple
string).  All r-package checks continue to pass.