Added functions to compare databases, formated code with styler
This commit adds some additional functions that can be used to compare
the structure and content of two bases (similar to db_diff).  It also
formatted the code with the styler package to ensure consistent coding
conventions throughout out the repository.  The package successfully
passes all checks and can be built successfully.