Added functions to plot proc-val errors. Added datetime utils
This commit adds some functions that create plots of FLEN, TLEN, RWT
and condition and super-imposes the appropriate limits specified in the
FN012 table. Additional reference fish can be added to the plots if
desired.  this commit also re-factors the data and time functionality
into separate functions that can be used to strip off the unnecessary
parts of timestamps when the fields are supposed to be either a date
or a time.  All of the existing tests continue to pass, but the new
functions are not document and the package will not currently build.