Update_readme function and fetch_table_data as is.
This commit replaces the populate_readme function with update readme
and changes that function to append rather than replace the readme in
the target database. It also updates the function call in
recode-database to fetch data as-is so that it does not drop leading
zeros from species and group codes. All tests continue to pass, and
the package builds without any warnings.